Make blast-stress-solver optional so Vercel can build shared wasm
Cargo eagerly resolves every path dependency during metadata
resolution, even when the target cfg or Cargo feature that gates it
is disabled.  That means cargo check / cargo metadata / wasm-pack
build all needed third_party/physx/blast/blast-stress-solver-rs/
to exist — which it doesn't on Vercel preview builds and other
environments that can't run scripts/setup-blast.sh.

Make blast-stress-solver an optional dep under a new `destructibles`
feature and drop a checked-in placeholder crate at
stubs/blast-stress-solver-rs/ that scripts/ensure-blast-stub.sh
copies into the third_party path when the real PhysX clone is
missing.  scripts/build-shared-wasm.sh auto-detects whether to pass
--features destructibles to wasm-pack.  shared/src/destructibles.rs
now re-exports either the real backend (destructibles_real.rs) or a
no-op stub backend (destructibles_stub.rs) based on the feature, so
wasm_api.rs compiles unchanged and the JS destructibles API still
exists on WasmSimWorld in both modes.

Verified: cargo check passes, wasm-pack build succeeds both with and
without third_party/physx/ present, all 339 client tests still pass,
tsc --noEmit clean.
