+++ AGENTS.md
@@ -116,7 +116,7 @@ Firewall requirements for WebTransport:
 
 - Rust toolchain must be >= 1.86. Run `rustup update stable && rustup default stable` if needed.
 - The shared WASM crate (`shared/`) is compiled separately with `wasm-pack` and output to `client/pkg/`. This must be rebuilt whenever `shared/` changes.
-- `make setup` now clones the pinned NVIDIA Blast stress solver into `third_party/physx/` (gitignored) via `scripts/setup-blast.sh`. The `blast-stress-solver` dep is wasm-only — `cargo check` / `cargo build -p web-fps-server` on a server dev box never touches the Blast C++ sources. See `docs/BLAST_INTEGRATION.md` for the full toolchain. The crate ships pure-Rust stubs for every libc symbol libc++ references, so the final wasm has zero `env.*` and zero `wasi_snapshot_preview1.*` imports and `make setup-wasm` is a single `wasm-pack build` with no post-processing.
+- `make setup` now clones the pinned NVIDIA Blast stress solver into `third_party/physx/` (gitignored) via `scripts/setup-blast.sh`. The `blast-stress-solver` dep is wasm-only AND optional (behind the `destructibles` Cargo feature) — `cargo check` / `cargo build -p web-fps-server` on a server dev box never touches the Blast C++ sources. Cargo still resolves the optional path dep during metadata, so a tiny placeholder crate at `stubs/blast-stress-solver-rs/` is dropped into `third_party/physx/blast/blast-stress-solver-rs/` by `scripts/ensure-blast-stub.sh` whenever the real clone is missing (e.g. Vercel preview builds). `scripts/build-shared-wasm.sh` detects whether the real PhysX clone is present and passes `--features destructibles` accordingly; when only the stub is present the wasm module still exposes the full destructibles JS API on `WasmSimWorld`, but every call is a no-op. See `docs/BLAST_INTEGRATION.md` for the full toolchain. The real crate ships pure-Rust stubs for every libc symbol libc++ references, so the final wasm has zero `env.*` and zero `wasi_snapshot_preview1.*` imports and `make setup-wasm` does no post-processing.
 - The web client is a single SPA build. Runtime route selection decides between `/play` multiplayer and `/practice` firing range; build mode no longer selects that behavior.
 - `VITE_MULTIPLAYER_HTTP_ORIGIN` is optional. Leave it unset for same-origin deployments; set it when the SPA is hosted separately from the Rust/WebTransport backend.
 - The `rapier3d` 0.30 API is used. Key notes: `KinematicCharacterController` is in `rapier3d::control`.
