+++ shared/Cargo.toml
@@ -28,12 +28,27 @@ serde_json = "1"
 default = []
 simd-stable = ["rapier3d/simd-stable", "vibe-netcode/simd-stable"]
 
+# Enables the real NVIDIA Blast stress solver backend for destructible
+# structures in the wasm build.  Requires a local clone of
+# `Glavin001/PhysX` at `third_party/physx/` (see
+# `scripts/setup-blast.sh`) AND a wasi C++ toolchain.  When this feature
+# is OFF the wasm module still exposes the destructibles JS API, but
+# every call is a no-op — useful for CI environments (e.g. Vercel
+# preview) that cannot build the C++ backend.
+destructibles = ["dep:blast-stress-solver"]
+
+# ─── Destructibles (Blast stress solver) ──────────────────────────────
+#
+# `blast-stress-solver` is an optional wasm-only dep.  Cargo still
+# needs a valid Cargo.toml at the path below during metadata
+# resolution, even when the `destructibles` feature is off — so
+# `scripts/ensure-blast-stub.sh` drops a tiny placeholder crate there
+# whenever `third_party/physx/` is missing (e.g. fresh clones, Vercel
+# preview).  On dev boxes `scripts/setup-blast.sh` clones the real
+# PhysX over the stub.  See `docs/BLAST_INTEGRATION.md` for the full
+# story.
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen = "0.2"
 js-sys = "0.3"
 console_error_panic_hook = "0.1"
-# blast-stress-solver is built from a local clone of Glavin001/PhysX produced
-# by `scripts/setup-blast.sh`. It is only compiled for wasm32 so native
-# `cargo check` / `cargo build -p vibe-land-server` on Linux developers
-# does not need a wasi C++ toolchain. See docs/BLAST_INTEGRATION.md.
-blast-stress-solver = { path = "../third_party/physx/blast/blast-stress-solver-rs", default-features = false, features = ["scenarios", "rapier"] }
+blast-stress-solver = { path = "../third_party/physx/blast/blast-stress-solver-rs", default-features = false, features = ["scenarios", "rapier"], optional = true }
